responsive layout updates
