+++ .gitignore
@@ -1,1 +1,2 @@
 config.js
+.DS_Store
