+++ contact_us.html
@@ -107,6 +107,7 @@
       <a id="facebook-link" href="https://www.facebook.com/people/Playful-Pathways-Therapy/100092405804453/?mibextid=LQQJ4d" target="_blank"><img  class = "facebook" alt = "facebook logo" src = "images/logo-facebook.png"></img></a>
       <p>©2023 by Playful Pathways Therapy, LLC.</p>
     </footer>
-    <script src="script.js" type="module"></script>
+    <script src="contact.js" type="module"></script>
+
   </body>
 </html>
