+++ styles.css
@@ -44,341 +44,64 @@ body {
     color: var(--second-color);
 }
 
-header a {
-    text-decoration: none;
-}
-
 header {
-    background-color: var(--main-color);
-    height: 50px;
-}
-
-nav ul {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.mobile-nav ul a,
-nav ul a {
-    color: var(--second-color);
-}
-
-.mobile-nav ul li,
-nav ul li {
-    padding: 10px;
-    margin-left: 10px;
-}
-
-.sub-menu {
-    display: none;
-    position: absolute;
-    top: 100%;
-    left: 0;
-    background-color: var(--main-color);
-    min-width: 150px;
-    border-radius: 5px;
-    text-align: center;
-}
-
-ul li:hover {
-    transform: scale(1.1);
-    transition: 0.3s;
-    font-weight: bold;
-}
-
-.mobile-nav ul li:hover .sub-menu,
-nav ul li:hover .sub-menu {
-    display: inherit;
-}
-
-#hamburger-icon {
-    margin: 0;
-    padding-top: 10px;
-    padding-left: 5px;
-    display: none;
-    cursor: pointer;
-}
-
-#hamburger-icon div {
-    width: 35px;
-    height: 3px;
-    background-color: var(--second-color);
-    margin: 6px 0;
-    transition: 0.4s;
-}
-
-.open .bar1 {
-    -webkit-transform: rotate(-45deg) translate(-6px, 6px);
-    transform: rotate(-45deg) translate(-6px, 6px);
-}
-
-.open .bar2 {
-    opacity: 0;
-}
-
-.open .bar3 {
-    -webkit-transform: rotate(45deg) translate(-6px, -8px);
-    transform: rotate(45deg) translate(-6px, -8px);
-}
-
-.open .mobile-menu {
-    display: flex;
-    flex-direction: column;
-    justify-content:flex-start;
-    align-items:center;
-}
-
-.mobile-menu {
-    display: none;
-    position: relative;
+    background-color: white;
     width: 100%;
-    height: 100%;
-    padding-bottom: 10px;
-    background-color: var(--main-color);
-}
-
-
-@media only screen and (max-width: 600px) {
-    nav {
-        display: none;
-    }
-    #hamburger-icon {
-        display: block;
-    }
+    height: auto;
 }
 
 .welcome-container {
     background-color: var(--main-color);
-    margin: 0 auto;
+}
+
+.welcome-box {
+    width: 80%;
+    max-width: 800px;
+    height: auto;
+    margin: auto;
+    margin-bottom: 40px;
+    text-align: center;
     padding-top: 20px;
-    padding-bottom: 80px;
+    padding-bottom: 20px;
 }
 
 .welcome-container h1 {
-    font-size: 60px;
-    font-weight: bolder;
+    font-weight: bold;
     margin: 20px auto 20px auto;
     color:  var(--heading-color);
 }
 
-.welcome-container p {
+.welcome-box p {
     line-height: 2em;
-    max-width:42%;
-    margin: 0 auto;
     padding-bottom: 20px;
 }
 
-@media only screen and (max-width: 600px) {
-    .welcome-container h1 {
-        font-size: 50px;
-    }
-
-    .welcome-container {
-        padding-top: 5px;
-        padding-bottom: 30px;
-    }
-    .welcome-container p {
-        max-width:70%;
-    }
-}
-
-.media-container {
-    overflow: hidden;
-    max-width: 100%;
-    white-space: nowrap;
-    height: auto;
-}
-
-.media-container img {
-    width: 35%;
-    height: auto;
-    max-width: 100%;
-    animation: scrollImages 15s ease-in-out infinite;
-    transition-delay: 10s;
-} 
-
-@keyframes scrollImages {
-    0% {
-        transform: translateX(0);
-    }
-    100% {
-        transform: translateX(calc(-100%));
-    }
-}
-
-@media only screen and (max-width: 600px) {
-    .media-container img {
-        width: 60%;
-        height: auto;
-    }
-}
-
 .practice-container {
-    display: flex;
-    flex-direction: row;
-    justify-content:center;
-    margin: 50px 80px;
-    height: auto;
-}
-
-.tableimg {
-    height: auto;
-    width: 50%;
-    max-width: 100%;
+    background-color: var(--heading-color);
 }
 
 .practice-scope-box {
-    width: 50%;
-    max-width: 100%;
-    height: auto;
-    background-color: var(--heading-color);
-    display: flex;
-}
-
-.practice-scope-text {
-    line-height: 1.5em;
-    max-width: 80%;
-    margin: auto;
     text-align: left;
     color: white;
 }
 
-.practice-scope-text p {
-    padding: 20px;
-}
-
-.practice-scope-text ul {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    padding-left: 20px;
-}
-
-@media only screen and (max-width: 600px) {
-    .practice-container {
-        flex-wrap: wrap;
-        flex-direction: column;
-        margin: 20px;
-    }
-
-    .tableimg,
-    .practice-scope-box {
-        width: 100%;
-        margin: unset;
-    }
-
-    .practice-scope-text {
-        max-width: 100%;
-        padding: 0 10px;
-    }
-}
-
 .slideshow-container {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    max-width: 100%;
-    height: auto;
-    min-height: 350px;
     background-color: rgb(248, 242, 248);
-}
-
-.containers {
-    display: none;
-}
-
-.back, .forward {
-    cursor: pointer;
-    position: absolute;
-    top: 46%;
-    width: auto;
-    margin-top: -22px;
-    padding: 0 75px;
-    color:  var(--heading-color);
-    font-size: 40px;
-    transition: 0.5s ease;
-    border-radius: 0 5px 5px 0;
-    user-select: none;
-  }
-
-  .forward {
-    right: 1rem;
-  }
-
-  .back {
-    left: 1rem;
-  }
-
-  .back:hover, .forward:hover {
+    padding: 40px;
     color: var(--second-color);
-  }
+}
 
-  .info {
-    text-align: center;
-    padding-top: 30px;
-  }
-
-  .quotation-mark {
-    font-size: 100px;
-    color: var(--main-color);
-    -webkit-text-stroke: 2px var(--second-color);
-    font-family: 'BIZ UDPGothic', sans-serif;
-    margin: 0;
-  }
-
-  .review {
-    max-width: 60%;
+.slideshow-container h2 {
+    color: var(--heading-color);
+    margin-bottom: 20px;
+}
+.review {
+    max-width: 70%;
     margin: auto;
-    font-size: 16px;
     font-style:italic;
-  }
-
-  .review-name {
-    font-size: 17px;
-
-  }
-
-  .dot-style {
-    text-align: center; 
-    padding-bottom: 20px; 
-    position: absolute; 
-    bottom: 5%; 
-    left: 48.5%;
-  }
-
-  @media only screen and (max-width: 600px) {
-    .slideshow-container {
-        height:500px;
-    }
-
-    .back, .forward {
-        padding: 0 10px;
-        font-size: 30px;
-    }
-
-    .dot-style {
-        left: 45%;
-    }
-  }
-
-  .dots {
-    cursor: pointer;
-    height: 8px;
-    width: 8px;
-    margin: 0 3px;
-    background-color: var(--second-color);
-    border-radius: 50%;
-    border: 2px solid var(--second-color);
-    display: inline-block;
-    transition: background-color 0.5s ease;
-  }
-
-  .enable, .dots:hover {
-    background-color: var(--main-color);
-    border-color: var(--second-color);
-  }
+    text-align: justify;
+    margin-bottom: 10px;
+}
 
 .hours-container {
     display: flex;
